refactor(step4): ui, domain 패키지 분리
