+++ src/main/kotlin/racing/domain/Cars.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 class Cars(
     private val cars: List<Car>
