+++ src/main/kotlin/racing/domain/Racing.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 data class Racing(
     val cars: Cars,
