+++ src/main/kotlin/racing/domain/RacingPhaseRecord.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 class RacingPhaseRecord(
     val distance: Int
