+++ src/main/kotlin/racing/domain/RacingPhaseRecords.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 class RacingPhaseRecords(
     val records: List<RacingPhaseRecord>
