+++ src/main/kotlin/racing/domain/RandomNumberForwardCondition.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 import kotlin.random.Random
