+++ src/main/kotlin/racing/main.kt
@@ -1,5 +1,11 @@
 package racing
 
+import racing.domain.Cars
+import racing.domain.Racing
+import racing.domain.RandomNumberForwardCondition
+import racing.ui.InputView
+import racing.ui.OutputView
+
 fun main() {
     val racingInputDto = InputView.read()
