+++ src/main/kotlin/racing/ui/InputView.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.ui
 
 object InputView {
