+++ src/main/kotlin/racing/ui/OutputView.kt
@@ -1,4 +1,6 @@
-package racing
+package racing.ui
+
+import racing.domain.RacingPhaseRecords
 
 object OutputView {
