+++ src/test/kotlin/racing/CarsTest.kt
@@ -2,6 +2,8 @@ package racing
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racing.domain.Cars
+import racing.domain.RandomNumberForwardCondition
 
 class CarsTest {
